Modify Package.targets to allow multiple packages
In the previous version, only a single package could be created and that was driven via the Input_Package* properties in the Input.targets file as well as a .nuspec.template file typically in your build directory.

Package creation in the new design is driven by the existance of a .nuspec.template file in the VS project director(ies) for which to build a package.  The scripts then follow the same approach of running 'nuget pack' with the .csproj but then using a .nuspec file (generated from the .nuspec.template) to feed in additional metadata.